Wrap Logger inside closure to use Logger of App we're macro'd into
When requiring Logger in MIX_ENV=prod environment, previous
implementation would fall over because of using different logger than
what is set in core application as logger backend.

This change uses the Logger that is in the macro context and passes it
in as an anonymous function to the Telemetry module.